+++ src/opcode.h
@@ -36,6 +36,7 @@ namespace ydsh {
   OP(PRINT, 3, -1)                                                                                 \
   OP(INSTANCE_OF, 3, 0)                                                                            \
   OP(CHECK_CAST, 3, 0)                                                                             \
+  OP(CHECK_CAST_OPT, 3, 0)                                                                         \
   OP(PUSH_NULL, 0, 1)                                                                              \
   OP(PUSH_TRUE, 0, 1)                                                                              \
   OP(PUSH_FALSE, 0, 1)                                                                             \